Fix crash on ending call.
We moved onUiShowing() call from InCallActivityCommon.onPause to onStop in
cl/169615685, which causes the crash since InCallPresenter is cleared up before
calling onUiShowing() and mContext is null at that time.

This CL fix the crash by also moving the clearup(InCallPresenter.unsetActivity()) from onPause
to onStop. It also makes sense because InCallPresenter.onActivityStarted() is
called in onStart instead of onPause.

Test: manual
PiperOrigin-RevId: 169756942
Change-Id: I222009be1b672f8a6dcb8bfdcea2d633c5810575