Use help center for GoogleDialer about
The about in settings is removed in google dialer, it is accessed through the dialtacts overflow "Help & feedback" instead:
https://screenshot.googleplex.com/9L1kdQz6c2B

https://screenshot.googleplex.com/FqqdG6X9JTR
Note that help articles has not been migrated at this moment.

Features in the original about is placed in the overflow
https://screenshot.googleplex.com/OmCVPqwGBM9

Version info is stock google help
https://screenshot.googleplex.com/EdveVnXUSML

Privacy policy/Terms of service opens the browser to the link we had before

Open source license opens the same license activity in about.

Bug: 34872689
Test: manual
PiperOrigin-RevId: 169953610
Change-Id: Ib1c25ddecbf28293b447112cb280a8f9788692d3