Group calls in the call log according to their corresponding actions.
Currently different types of calls can be grouped together (e.g., 1 voice call and 1 IMS call to/from the same number), which makes it difficult to choose the icon for the call detail UI's call back button.

This CL adds an extra constraint that separates different call types (Lightbringer, IMS, and voice). This way calls in the call detail UI are in the same category and an appropriate icon can be set.

Bug: 66026167
Test: CallLogGroupBuilderTest.addGroups_MixedEntries_PartiallyGroupedByAction
PiperOrigin-RevId: 171602617
Change-Id: Id8170206009ba836a40c38a86914c71d5c7701dc