Do not load every DialtactsActivity fragment on app open
With this change we now only load the Favorites and Call Log Fragments when the
app opens.

Paging to the CallLogFragment makes the ContactsFragment load as we would want.

This should result in a faster app start time. I'll need to verify the impact
of this using the same method as http://cl/170422671

I verified the change in behavior by adding my own log line to the onCreate of each Fragment. I saw all of them get created before the change, and only the expected 2 with this change.

Bug: 64541209,65460373
Test: manual
PiperOrigin-RevId: 172154863
Change-Id: Iaf6e217df284e6df6f13688ac6aa677481a575e0