Fixed crash which occurred when clearing call log containing enriched call data.
There was a race where the dialog could sometimes get detached before the success callback finished. The callback would attempt to show a snackbar using [fragment.]getString(), which throws when the fragment is detached.

The fix is to just use the activity to get the string (which has already been asserted to be valid).

Bug: 65968151
Test: unit
PiperOrigin-RevId: 172169503
Change-Id: If86250b14038958e6668b26d1f150619fe3eefae