Display Voicemail Contact Name and Photo in Voicemail Fragment.
This CL lays the foundation of querying the annotated call log, retrieving only the voicemail entries and some of the voicemail information in the Voicemail tab. This CL only displays the name/number as well as the corresponding photo of the voicemails present on the device.

WANT_LGTM=zachh
SCREENSHOT: http://screen/YYBG8T0Ma0J
Bug: 33006245, 64882313
Test: Unit tests and manual verification.
PiperOrigin-RevId: 172504512
Change-Id: Ie7187b021abf42d1a4522458a1ab28537a7c222c