Support bulk inserted voicemails.
Bulk insert is implemented in ag/3073902

When more than one voicemail is inserted into the VoicemailProvider, the URI type is Voicemails.DIR_TYPE instead of ITEM_TYPE. This CL adds support to the new type.

Previously when multiple voicemails are being notified, each one of them will generate a ringtone because the notifiction is set to alert children. In this CL, the summary will have alert instead, and will only alert when new voicemails are received.

Bug: 21281858
Test: VisualVoicemailNotifierTest
PiperOrigin-RevId: 172546076
Change-Id: I4df6098fc3a793139e3d4e77fe1339efdc001944