Take into account special dialable characters when grouping calls in the call log UI.
Currently we use PhoneNumberUtils#compare(String, String) to determine whether two numbers in the call log should be grouped. The method ignores special dialable characters such as "#", "*", "+", etc, which can cause phone numbers and service dialing numbers to be grouped.

For example, suppose the user has a contact named "John Smith" with number 123456789 and a contact named "Service" with number #123456789, and called each contact once. The old grouping logic will put the two numbers in the same group.

The new grouping logic in this CL can separate the two numbers.

Bug: 30225112
Test: CallLogGroupBuilderTest, PhoneNumberHelperTest
PiperOrigin-RevId: 172683494
Change-Id: Ie9e7b7418d6d7c74830cbae09c04ff1feb1b827d