Automated g4 rollback of changelist 172154863.
*** Reason for rollback ***

Causes crashes on rotation: http://b/67839843

*** Original change description ***

Do not load every DialtactsActivity fragment on app open

With this change we now only load the Favorites and Call Log Fragments when the
app opens.

Paging to the CallLogFragment makes the ContactsFragment load as we would want.

This should result in a faster app start time. I'll need to verify the impact
of this using the same method as http://cl/170422671

I verified the change in behavior by adding my own log line to the onCreate of each Fragment. I saw all of them get created before the ch...

***

Bug: 64541209,65460373
Test: Manually verified regression fixed
PiperOrigin-RevId: 172684553
Change-Id: I9a5b30d6619d930adfb82b19ac3ff9da00decaaa