Fix HOLD option not displayed in voice calls in CDMA + GSM.
Add hold button for all phone types and rely on SUPPORT_HOLD capability to show it.

The issue is that if GSM call is made on devices with CDMA + GSM sim, the default phone type is CDMA so the hold button is not available. This change is a workaround that adding hold button to CDMA button chooser so button choose of all phone type has hold button available and it's totally controlled by telecom's CAPABILITY_SUPPORT_HOLD and CAPABILITY_HOLD. It shouldn't affect CDMA call.

Idealy we should use correct phone type but unfortunately it's not well supported on pre-O device.

This is upstream change from
https://android-review.googlesource.com/#/c/platform/packages/apps/Dialer/+/482377/

Bug: 63911282
Test: ButtonChooserIntegrationTest
PiperOrigin-RevId: 173614690
Change-Id: I6e17c1cc69ad50b1ba4c9b09804d5a7b5eeaf496