Force disable VVM if carrier app is preloaded
This CL checks if the carrier app is installed on the system image. If so, VVM will be force disabled (user can't turn it on, can't even see the switch as if they are on a unsupported carrier). Otherwise, VVM will only be disabled by default if the carrier app is present (User can still turn it on in settings).

Bug: 67496592
Test: OmtpCarrierConfigHelperTest / manual - move carrier app to /system/app
PiperOrigin-RevId: 176586849
Change-Id: Idf79bdf61c600423062ca54d1b4c6e77cfcd8033