Play voicemails, update seekbar timer, allow seeking, and maintain state after recycling.
This CL adds the support for playing voicemails, changing the play button to pause button when playing the voicemail, updating the seekbar and duration timer when the voicemail is being played. It also adds the support to preserve the state of the media player such that when scrolling and recycling views, when an expanded and playing voicemail is recycled back into view, it's most recent state is shown i.e the duration and the seekbar are upto date.

Video: https://drive.google.com/open?id=1CKbLK5-1YDeXBZFiKvuTxoPuFJQ1rbj7
Test: Unit tests
PiperOrigin-RevId: 178681663
Change-Id: Ifdd1d945572926bdc7d652aa7a876d3156fc21ce