Only consider writable accounts for preferred SIM
Some apps will add their own account to contacts and cause PhoneLookup to return multiple accounts. Since Contacts cannot write to these accounts the preferred SIM settings on them does not make sense.

In this CL a reduced version of AccountTypeManager is used to determine if the account is writable. External accounts (which capabilities are determined by parsing their sync adapter) are not supported.

Bug: 70689051
Test: CallingAccountSelectorTest
PiperOrigin-RevId: 180693656
Change-Id: I24e93860cd576777c6d3861f65f75baa234dac87