Make sure that voicemail transcriptions are processed serially
Since we are now using the AlarmManager to decide when to poll
for voicemail transcriptions, we need to ensure that we don't
start transcribing a voicemail until polling for the previous
one ends.

This cl postpones a voicemail transcription task if there is an alarm
set to poll for a transcription result.  After a transcription result
is received (or we give up) the db is scanned for any pending transcriptions
and the next one is started.

Bug: 70242961
Test: manual and updated unit tests
PiperOrigin-RevId: 181899975
Change-Id: I7b8fb696164980cf710aa58a79418c6954e2b4d2