Relax number matching when coalescing rows in new call log.
We currently require two numbers to be a PhoneNumberUtil.MatchType.EXACT_MATCH to coalesce them, but this means that two numbers like "456-7890" and "408 456-7890" won't ever be collapsed.

This is potentially a likely situation since it is possible to dial numbers without an area code so we should better support it (and the old call log coalesces such numbers today).

Bug: 70989626
Test: unit
PiperOrigin-RevId: 182289194
Change-Id: If884d5a1f2631116a2729e0635f9a97aeca3e057