Route speaker via speaker and headset when tapping speaker button in NUI VM
Clicking the speaker button in the NUI VM for a voicemail entry alternates playing the audio via the speaker and headset. This CL does not change the state of the speaker icon, that will be done in a follow up CL (cl/183175378) and the todo for that has been added.

Bug: 69858226
Test: N/A
PiperOrigin-RevId: 183304029
Change-Id: I3a55e62ab8d1203f659a45db9534fd69deda122f