Refresh annotated call log from the voicemail tab
This CL copies the code dealing with refreshing the annotated call log from the NewCallLogFragment and applies it to the NewVoicemailFragment. This is necessary since the NewVoicemailFragment should be able to refresh the annotated call log without depending on the user going to the call log fragment for the update. The only thing we don't copy over is the logic and handling when the CP2 info is incomplete for a number (INVALID number), however we do not expect any invalid numbers for voicemail and as such have added an assert for that.

Bug: 72523272
Test: Unit Tests
PiperOrigin-RevId: 184292549
Change-Id: I2994974fe60221a4854fed5f9d80d3d428449ae9